firmware: Add NSS EIP firmware package
This is just a helper package that helps ensure
that the EIP-197 firmware is installed by being
used as a dependency on nss-crypto.

Note that the package does not provide or ship
with firmware binaries itself.
Those need to be pulled from the stock firmware.
Binaries required are:
* ifpp.bin
* ipue.bin
* ofpp.bin
* opue.bin

They should be placed inside of the src folder.

Note that we add the src folder to the .gitignore to
prevent accidentally pushing the FW to a public repo.

Signed-off-by: Robert Marko <robimarko@gmail.com>